Move electric cleaner into namespace
