--- TombEngine/Objects/TR3/Trap/cleaner.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "Game/items.h"
-#include "Game/collision/collide_room.h"
-
-void InitialiseCleaner(short itemNumber);
-void CleanerControl(short itemNumber);
-void CleanerCollision(short itemNumber, ItemInfo* laraItem, CollisionInfo* coll);
